Refactor package namespace structure
