--- contracts/LikeCounter.php
@@ -1,29 +1,0 @@
-<?php
-
-/*
- * This file is part of Laravel Likeable.
- *
- * (c) Anton Komarev <a.komarev@cybercog.su>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Cog\Contracts\Likeable;
-
-/**
- * Interface LikeCounter.
- *
- * @property int type_id
- * @property int count
- * @package Cog\Contracts\Likeable
- */
-interface LikeCounter
-{
-    /**
-     * Likeable model relation.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphTo
-     */
-    public function likeable();
-}
