--- src/Enums/LikeType.php
@@ -1,23 +1,0 @@
-<?php
-
-/*
- * This file is part of Laravel Likeable.
- *
- * (c) Anton Komarev <a.komarev@cybercog.su>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Cog\Laravel\Likeable\Enums;
-
-/**
- * Class LikeType.
- *
- * @package Cog\Laravel\Likeable\Enums
- */
-class LikeType
-{
-    const LIKE = 'like';
-    const DISLIKE = 'dislike';
-}
