+++ src/Like/Enums/LikeType.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of Laravel Likeable.
+ *
+ * (c) Anton Komarev <a.komarev@cybercog.su>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cog\Laravel\Likeable\Like\Enums;
+
+/**
+ * Class LikeType.
+ *
+ * @package Cog\Laravel\Likeable\Like\Enums
+ */
+class LikeType
+{
+    const LIKE = 'like';
+    const DISLIKE = 'dislike';
+}
