+++ tests/Unit/Like/Observers/LikeObserverTest.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of Laravel Likeable.
+ *
+ * (c) Anton Komarev <a.komarev@cybercog.su>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cog\Tests\Laravel\Likeable\Unit\Like\Observers;
+
+use Cog\Tests\Laravel\Likeable\TestCase;
+use Illuminate\Foundation\Testing\DatabaseTransactions;
+
+/**
+ * Class LikeObserverTest.
+ *
+ * @package Cog\Tests\Laravel\Likeable\Unit\Like\Observers
+ */
+class LikeObserverTest extends TestCase
+{
+    use DatabaseTransactions;
+}
