+++ tests/Unit/Services/LikeableServiceTest.php
@@ -11,8 +11,8 @@
 
 namespace Cog\Tests\Laravel\Likeable\Unit\Services;
 
-use Cog\Laravel\Likeable\Models\LikeCounter;
-use Cog\Contracts\Likeable\LikeableService as LikeableServiceContract;
+use Cog\Laravel\Likeable\LikeCounter\Models\LikeCounter;
+use Cog\Contracts\Likeable\Services\LikeableService as LikeableServiceContract;
 use Cog\Laravel\Likeable\Services\LikeableService;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Article;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
